Restore existing xattr allocation cast

+++ services/clouddiskservice/ipc/src/cloud_disk_service.cpp
@@ -537,7 +537,7 @@ static ResultList GetFileSyncState(const std::string &path, int32_t &userId, con
         return getResult;
     }
 
-    std::unique_ptr<char[]> xattrValue = std::make_unique<char[]>(static_cast<size_t>(xattrValueSize));
+    std::unique_ptr<char[]> xattrValue = std::make_unique<char[]>((long)xattrValueSize);
     if (xattrValue == nullptr) {
         LOGE("Failed to allocate memory for xattrValue, errno : %{public}d", errno);
         getResult.error = ErrorReason::NO_SPACE_LEFT;
